# Properly calculate compressed message lengths