--- a/dns.go
+++ b/dns.go
@@ -36,6 +36,8 @@ type RR interface {
 	copy() RR
 	// len returns the length (in octets) of the uncompressed RR in wire format.
 	len() int
+	// compressedLen returns the length (in octets) of the compressed RR in wire format.
+	compressedLen(off int, compression map[string]struct{}, compress bool) int
 	// pack packs an RR into wire format.
 	pack([]byte, int, map[string]int, bool) (int, error)
 }
@@ -72,6 +74,12 @@ func (h *RR_Header) String() string {
 
 func (h *RR_Header) len() int {
 	l := len(h.Name) + 1
+	l += 10 // rrtype(2) + class(2) + ttl(4) + rdlength(2)
+	return l
+}
+
+func (h *RR_Header) compressedLen(off int, compression map[string]struct{}, compress bool) int {
+	l := compressedNameLen(h.Name, off, compression, compress)
 	l += 10 // rrtype(2) + class(2) + ttl(4) + rdlength(2)
 	return l
 }