--- a/dns.go
+++ b/dns.go
@@ -86,7 +86,7 @@ func (h *RR_Header) compressedLen(off int, compression map[string]struct{}, comp
 
 // ToRFC3597 converts a known RR to the unknown RR representation from RFC 3597.
 func (rr *RFC3597) ToRFC3597(r RR) error {
-	buf := make([]byte, r.len()*2)
+	buf := make([]byte, Len(r)*2)
 	off, err := PackRR(r, buf, 0, nil, false)
 	if err != nil {
 		return err