--- a/edns.go
+++ b/edns.go
@@ -78,18 +78,8 @@ func (rr *OPT) String() string {
 	return s
 }
 
-func (rr *OPT) len() int {
-	l := rr.Hdr.len()
-	for i := 0; i < len(rr.Option); i++ {
-		l += 4 // Account for 2-byte option code and 2-byte option length.
-		lo, _ := rr.Option[i].pack()
-		l += len(lo)
-	}
-	return l
-}
-
-func (rr *OPT) compressedLen(off int, compression map[string]struct{}, compress bool) int {
-	l := rr.Hdr.compressedLen(off, compression, compress)
+func (rr *OPT) len(off int, compression map[string]struct{}, compress bool) int {
+	l := rr.Hdr.len(off, compression, compress)
 	for i := 0; i < len(rr.Option); i++ {
 		l += 4 // Account for 2-byte option code and 2-byte option length.
 		lo, _ := rr.Option[i].pack()