--- a/msg.go
+++ b/msg.go
@@ -923,21 +923,21 @@ func compressedLenWithCompressionMap(dns *Msg, compression map[string]struct{}, 
 	l := 12 // Message header is always 12 bytes
 
 	for _, r := range dns.Question {
-		l += r.compressedLen(l, compression, compress)
+		l += r.len(l, compression, compress)
 	}
 	for _, r := range dns.Answer {
 		if r != nil {
-			l += r.compressedLen(l, compression, compress)
+			l += r.len(l, compression, compress)
 		}
 	}
 	for _, r := range dns.Ns {
 		if r != nil {
-			l += r.compressedLen(l, compression, compress)
+			l += r.len(l, compression, compress)
 		}
 	}
 	for _, r := range dns.Extra {
 		if r != nil {
-			l += r.compressedLen(l, compression, compress)
+			l += r.len(l, compression, compress)
 		}
 	}
 
@@ -996,7 +996,7 @@ func compressionLenMapInsert(c map[string]struct{}, s string, msgOff int) {
 func Copy(r RR) RR { r1 := r.copy(); return r1 }
 
 // Len returns the length (in octets) of the uncompressed RR in wire format.
-func Len(r RR) int { return r.len() }
+func Len(r RR) int { return r.len(0, nil, false) }
 
 // Copy returns a new *Msg which is a deep-copy of dns.
 func (dns *Msg) Copy() *Msg { return dns.CopyTo(new(Msg)) }