This entirely replaces the rather hacky and very confusing dance done in `compressionLenSlice`, with a new approach that is far more like how packing is performed. It fixes a number of different bugs (some listed at the end).

The new code has two allocations that aren't actually needed, but are hard to avoid. (`MsgLength` & `MsgLengthMassive` should have 0 and 9 allocations respectively). The problem is that `compression map[string]struct{}` is passed to `RR.len` where `RR` is an interface. This interacts badly with Golang's escape analysis (see golang/go#19361 and <https://npat-efault.github.io/programming/2016/10/10/escape-analysis-and-interfaces.html>), forcing the map to be heap allocated instead of stack allocated. It is possible to force these to be devirtualised by using a massive type switch statement, this worsens performance overall for `MsgLengthMassive` while eliminating the loss for `MsgLength`.

The performance loss in `MsgLengthNoCompression` & `MsgLengthOnlyQuestion` is (I believe) due to the call to `domainNameLen` which can't be inlined currently. Though this is unfortunate, the overhead is necessary as it fixes a bug where `Len` could previously be off by one for uncompressed messages. (See the existing `TestMsgLength2` that is now fixed). Inlining changes are coming to go1.12 so this can be revisited then and it may be possible to recover this performance loss.

Some of the benchmarks below (like `LenLabels` and `MuxMatch`) are just noise (although I don't quite understand how they survived through benchstat and ten runs), but they could come down to the compiler laying out code differently and cache interactions. TLDR; they're a distraction.

When taken along with #820, `MsgLengthMassive` is now twice as fast, with half the allocated bytes and 92% fewer allocations.

<details>
<summary>$ benchstat {old,new}.bench</summary>
<pre>
$ benchstat {old,new}.bench
name                            old time/op    new time/op    delta
MsgLength-12                       336ns ± 0%     545ns ± 6%   +62.24%  (p=0.001 n=6+8)
MsgLengthNoCompression-12         9.95ns ± 3%   26.64ns ± 2%  +167.79%  (p=0.000 n=10+10)
MsgLengthPack-12                  2.20µs ±15%    2.30µs ±23%      ~     (p=0.315 n=10+10)
MsgLengthMassive-12               54.0µs ± 5%    31.8µs ± 6%   -41.04%  (p=0.000 n=10+9)
MsgLengthOnlyQuestion-12          5.53ns ± 1%    9.19ns ± 1%   +66.15%  (p=0.000 n=10+10)
PackDomainName-12                  211ns ± 6%     204ns ± 7%      ~     (p=0.064 n=10+10)
UnpackDomainName-12                150ns ±14%     157ns ±14%      ~     (p=0.383 n=10+10)
UnpackDomainNameUnprintable-12     145ns ± 3%     150ns ± 4%      ~     (p=0.089 n=9+10)
Copy-12                            880ns ±20%     742ns ±55%      ~     (p=0.247 n=10+10)
PackA-12                          41.1ns ± 2%    40.9ns ± 1%    -0.56%  (p=0.044 n=10+9)
UnpackA-12                         222ns ±21%     208ns ±41%      ~     (p=0.726 n=10+10)
PackMX-12                         75.1ns ± 1%    73.4ns ± 1%    -2.24%  (p=0.000 n=10+10)
UnpackMX-12                        278ns ± 8%     278ns ±16%      ~     (p=0.913 n=8+10)
PackAAAAA-12                      40.6ns ± 2%    40.2ns ± 4%      ~     (p=0.055 n=10+10)
UnpackAAAA-12                      243ns ±12%     230ns ± 9%      ~     (p=0.129 n=9+7)
PackMsg-12                        1.57µs ± 8%    1.57µs ±14%      ~     (p=0.853 n=10+10)
PackMsgOnlyQuestion-12             234ns ± 1%     237ns ± 2%    +1.14%  (p=0.046 n=9+9)
UnpackMsg-12                      1.38µs ±29%    1.37µs ±19%      ~     (p=0.853 n=10+10)
IdGeneration-12                   15.7ns ± 2%    15.7ns ± 0%      ~     (p=0.666 n=10+7)
Generate-12                        153µs ± 2%     153µs ± 1%      ~     (p=0.739 n=10+10)
SplitLabels-12                    69.6ns ±12%    62.0ns ±29%      ~     (p=0.063 n=10+10)
LenLabels-12                      27.2ns ± 1%    20.1ns ± 1%   -25.96%  (p=0.000 n=10+8)
CompareDomainName-12               166ns ±17%     161ns ±11%      ~     (p=0.343 n=10+10)
IsSubDomain-12                     529ns ± 3%     493ns ±20%      ~     (p=0.327 n=8+10)
UnpackString-12                    137ns ± 6%     131ns ± 9%      ~     (p=0.089 n=9+10)
Dedup-12                          1.94µs ± 8%    1.86µs ±11%      ~     (p=0.128 n=10+10)
NewRR-12                          2.64µs ± 5%    2.62µs ± 5%      ~     (p=0.289 n=10+10)
ReadRR-12                         4.28µs ±20%    4.10µs ±19%      ~     (p=0.684 n=10+10)
ParseZone-12                       123µs ±14%     115µs ±48%      ~     (p=0.971 n=10+10)
ZoneParser-12                     10.4µs ± 1%    10.6µs ± 4%      ~     (p=0.243 n=9+10)
MuxMatch/lowercase-12             69.7ns ± 2%    67.9ns ± 2%    -2.65%  (p=0.000 n=10+10)
MuxMatch/uppercase-12              120ns ± 4%     125ns ± 3%    +4.26%  (p=0.000 n=10+10)
Serve-12                          56.5µs ± 8%    56.4µs ± 4%      ~     (p=0.853 n=10+10)
Serve6-12                         57.3µs ± 5%    57.7µs ± 5%      ~     (p=0.684 n=10+10)
ServeCompress-12                  57.8µs ± 2%    58.5µs ± 3%      ~     (p=0.800 n=2+3)
SprintName-12                      218ns ± 2%     206ns ± 2%    -5.42%  (p=0.000 n=10+10)
SprintTxtOctet-12                  234ns ± 5%     233ns ±12%      ~     (p=0.858 n=9+10)
SprintTxt-12                       267ns ± 8%     277ns ± 7%      ~     (p=0.147 n=10+10)
&nbsp;
name                            old alloc/op   new alloc/op   delta
MsgLength-12                       32.0B ± 0%    192.0B ± 0%  +500.00%  (p=0.000 n=10+10)
MsgLengthNoCompression-12          0.00B          0.00B           ~     (all equal)
MsgLengthPack-12                    896B ± 0%      896B ± 0%      ~     (all equal)
MsgLengthMassive-12               14.8kB ± 0%    10.9kB ± 0%   -26.59%  (p=0.000 n=10+10)
MsgLengthOnlyQuestion-12           0.00B          0.00B           ~     (all equal)
PackDomainName-12                  64.0B ± 0%     64.0B ± 0%      ~     (all equal)
UnpackDomainName-12                64.0B ± 0%     64.0B ± 0%      ~     (all equal)
UnpackDomainNameUnprintable-12     48.0B ± 0%     48.0B ± 0%      ~     (all equal)
Copy-12                             432B ± 0%      432B ± 0%      ~     (all equal)
PackA-12                           0.00B          0.00B           ~     (all equal)
UnpackA-12                          100B ± 0%      100B ± 0%      ~     (all equal)
PackMX-12                          0.00B          0.00B           ~     (all equal)
UnpackMX-12                         116B ± 0%      116B ± 0%      ~     (all equal)
PackAAAAA-12                       0.00B          0.00B           ~     (all equal)
UnpackAAAA-12                       100B ± 0%      100B ± 0%      ~     (all equal)
PackMsg-12                          576B ± 0%      576B ± 0%      ~     (all equal)
PackMsgOnlyQuestion-12             64.0B ± 0%     64.0B ± 0%      ~     (all equal)
UnpackMsg-12                        592B ± 0%      592B ± 0%      ~     (all equal)
IdGeneration-12                    0.00B          0.00B           ~     (all equal)
Generate-12                       31.9kB ± 0%    31.9kB ± 0%      ~     (p=1.000 n=10+10)
SplitLabels-12                     32.0B ± 0%     32.0B ± 0%      ~     (all equal)
LenLabels-12                       0.00B          0.00B           ~     (all equal)
CompareDomainName-12               64.0B ± 0%     64.0B ± 0%      ~     (all equal)
IsSubDomain-12                      192B ± 0%      192B ± 0%      ~     (all equal)
UnpackString-12                    48.0B ± 0%     48.0B ± 0%      ~     (all equal)
Dedup-12                            624B ± 0%      624B ± 0%      ~     (all equal)
NewRR-12                            784B ± 0%      784B ± 0%      ~     (all equal)
ReadRR-12                         1.79kB ± 0%    1.79kB ± 0%      ~     (all equal)
ParseZone-12                      84.0kB ± 0%    84.0kB ± 0%      ~     (all equal)
ZoneParser-12                     1.57kB ± 0%    1.57kB ± 0%      ~     (all equal)
MuxMatch/lowercase-12              0.00B          0.00B           ~     (all equal)
MuxMatch/uppercase-12              32.0B ± 0%     32.0B ± 0%      ~     (all equal)
Serve-12                          3.36kB ± 0%    3.36kB ± 0%      ~     (all equal)
Serve6-12                         3.20kB ± 0%    3.20kB ± 0%      ~     (all equal)
ServeCompress-12                  3.62kB ± 0%    3.62kB ± 0%      ~     (all equal)
SprintName-12                      48.0B ± 0%     48.0B ± 0%      ~     (all equal)
SprintTxtOctet-12                  80.0B ± 0%     80.0B ± 0%      ~     (all equal)
SprintTxt-12                       80.0B ± 0%     80.0B ± 0%      ~     (all equal)
&nbsp;
name                            old allocs/op  new allocs/op  delta
MsgLength-12                        1.00 ± 0%      2.00 ± 0%  +100.00%  (p=0.000 n=10+10)
MsgLengthNoCompression-12           0.00           0.00           ~     (all equal)
MsgLengthPack-12                    8.00 ± 0%      8.00 ± 0%      ~     (all equal)
MsgLengthMassive-12                  138 ± 0%        11 ± 0%   -92.03%  (p=0.000 n=10+10)
MsgLengthOnlyQuestion-12            0.00           0.00           ~     (all equal)
PackDomainName-12                   1.00 ± 0%      1.00 ± 0%      ~     (all equal)
UnpackDomainName-12                 1.00 ± 0%      1.00 ± 0%      ~     (all equal)
UnpackDomainNameUnprintable-12      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
Copy-12                             8.00 ± 0%      8.00 ± 0%      ~     (all equal)
PackA-12                            0.00           0.00           ~     (all equal)
UnpackA-12                          3.00 ± 0%      3.00 ± 0%      ~     (all equal)
PackMX-12                           0.00           0.00           ~     (all equal)
UnpackMX-12                         4.00 ± 0%      4.00 ± 0%      ~     (all equal)
PackAAAAA-12                        0.00           0.00           ~     (all equal)
UnpackAAAA-12                       3.00 ± 0%      3.00 ± 0%      ~     (all equal)
PackMsg-12                          7.00 ± 0%      7.00 ± 0%      ~     (all equal)
PackMsgOnlyQuestion-12              1.00 ± 0%      1.00 ± 0%      ~     (all equal)
UnpackMsg-12                        12.0 ± 0%      12.0 ± 0%      ~     (all equal)
IdGeneration-12                     0.00           0.00           ~     (all equal)
Generate-12                        1.55k ± 0%     1.55k ± 0%      ~     (all equal)
SplitLabels-12                      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
LenLabels-12                        0.00           0.00           ~     (all equal)
CompareDomainName-12                2.00 ± 0%      2.00 ± 0%      ~     (all equal)
IsSubDomain-12                      6.00 ± 0%      6.00 ± 0%      ~     (all equal)
UnpackString-12                     2.00 ± 0%      2.00 ± 0%      ~     (all equal)
Dedup-12                            31.0 ± 0%      31.0 ± 0%      ~     (all equal)
NewRR-12                            15.0 ± 0%      15.0 ± 0%      ~     (all equal)
ReadRR-12                           17.0 ± 0%      17.0 ± 0%      ~     (all equal)
ParseZone-12                        92.0 ± 0%      92.0 ± 0%      ~     (all equal)
ZoneParser-12                       81.0 ± 0%      81.0 ± 0%      ~     (all equal)
MuxMatch/lowercase-12               0.00           0.00           ~     (all equal)
MuxMatch/uppercase-12               1.00 ± 0%      1.00 ± 0%      ~     (all equal)
Serve-12                            54.0 ± 0%      54.0 ± 0%      ~     (all equal)
Serve6-12                           51.0 ± 0%      51.0 ± 0%      ~     (all equal)
ServeCompress-12                    56.0 ± 0%      56.0 ± 0%      ~     (all equal)
SprintName-12                       2.00 ± 0%      2.00 ± 0%      ~     (all equal)
SprintTxtOctet-12                   2.00 ± 0%      2.00 ± 0%      ~     (all equal)
SprintTxt-12                        2.00 ± 0%      2.00 ± 0%      ~     (all equal)
</pre>
</details><br>

Updates #709
Fixes #821
Fixes #824
Closes #826
Fixes #829

/cc @pierresouchay (who changed much of this in #668).