@miekg 2018-11-28
[ Quoting <notifications@github.com> in "Re: [miekg/dns] Properly calculate ..." ]
>I had a quick look, it seems reasonable, and removing my crappy hacks is good news ;)
>I'll try to bench the performance effects in Consul once merged

Why not before? ;-)

/Miek

--
Miek Gieben